Import CSV Files into Objects
This main looks for csv files in the runtime arguments. For csv files found in argument list, it imports the csv as a matrix object into a circular linked list of objects. Correlations are made between the csv file data property values and all other csv files. Then this main displays objects identified during object import based on the data in the csv files found in arguments during runtime.